+++ examples/hello_world/fabfile.py
@@ -6,7 +6,7 @@ nginx = tasks.DockerTasks(
         name='nginx',
         image='nginx:stable-alpine',
         options=dict(
-            ports='80:80',
+            publish='80:80',
         ),
     ),
     hosts=AvailableVagrantHosts(),